reorganize test fixtures
### Description
Just moving things around.  This doesn't actually change any code.

### Test plan
pytest FTW?
