+++ src/py/aspen/conftest.py
@@ -1,15 +1,11 @@
-from aspen.test_app.fixtures import app, client  # noqa: F401
-from aspen.test_infra.fixtures import (  # noqa: F401
+from aspen.test_infra.flask import app, client  # noqa: F401
+from aspen.test_infra.models.sample import sample  # noqa: F401
+from aspen.test_infra.models.sequences import sequencing_read  # noqa: F401
+from aspen.test_infra.models.usergroup import group, user  # noqa: F401
+from aspen.test_infra.postgres import (  # noqa: F401
     postgres_database,
     postgres_database_with_schema,
     postgres_instance,
-    session,
-    sqlalchemy_interface,
     unused_tcp_port,
 )
-from aspen.test_models.fixtures import (  # noqa: F401
-    group,
-    sample,
-    sequencing_read,
-    user,
-)
+from aspen.test_infra.sqlalchemy import session, sqlalchemy_interface  # noqa: F401
