+++ src/py/aspen/test_infra/postgres.py
@@ -7,7 +7,6 @@ from typing import Generator
 from uuid import uuid4
 
 import pytest
-from sqlalchemy.orm import Session
 
 from aspen.database import connection as aspen_connection
 from aspen.database import schema
@@ -178,24 +177,3 @@ def postgres_database_with_schema(
     db = aspen_connection.init_db(postgres_database.as_uri())
     schema.create_tables_and_schema(db)
     yield postgres_database
-
-
-@pytest.fixture()
-def sqlalchemy_interface(
-    postgres_database_with_schema,
-) -> Generator[aspen_connection.SqlAlchemyInterface, None, None]:
-    interface = aspen_connection.init_db(postgres_database_with_schema.as_uri())
-    connection = interface.engine.connect()
-
-    yield interface
-
-    connection.close()
-
-
-@pytest.fixture()
-def session(sqlalchemy_interface) -> Generator[Session, None, None]:
-    session = sqlalchemy_interface.make_session()
-
-    yield session
-
-    session.close()
